requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.7
django-crispy-forms==2.1
django-multiupload==0.6.1
django-shortuuidfield==0.1.3
environs==9.5.0
marshmallow==3.20.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.8
python-dotenv==1.0.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
